+++ packages/server/src/http/index.ts
@@ -2,7 +2,14 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { assertNotBrowser } from '../assertNotBrowser';
 
-export * from './requestHandler';
+// @deprecated delete in next major
+export { requestHandler } from '../adapters/node-http/requestHandler';
+// @deprecated delete in next major
+export type {
+  CreateContextFn,
+  CreateContextFnOptions,
+} from '../adapters/node-http/types';
+
 export * from './ResponseMeta';
 
 assertNotBrowser();
